received first completion from api
