--- api.py
@@ -1,17 +1,0 @@
-import requests
-
-from init.tokens import hf_token
-
-API_URL = "https://api-inference.huggingface.co/models/mistralai/Mistral-7B-v0.1"
-headers = {"Authorization": f"Bearer {hf_token}"}
-
-def query(payload):
-	response = requests.post(API_URL, headers=headers, json=payload)
-	return response.json()
-	
-output = query({
-	"inputs": "Can you please let us know more details about your ",
-})
-
-print(output)
-print("hey")
